Split out assertions where NPEs can occur in TestPKITest
IntelliJ complains that NPEs can be thrown by both getHandshake()
and peerPrincipal() methods in the expression:

recordedRequest.getHandshake().peerPrincipal().toString()

Since this library uses Java 8, if an NPE is thrown, you won't get a
nice error message explaining which object was null, like you do
in Java 14 and higher. This splits out the assertions to verify
each object, which is more verbose, but also much easier to determine
where the problem is (should one occur). It's not as concise, but
provides more useful information when there is a problem.